rafctor: move MeshNode struct to the separate file
