+++ src/graphics/vulkan/vk_engine.cpp
@@ -1395,25 +1395,6 @@ MaterialInstance GLTFMetallic_Roughness::write_material(
     return matData;
 }
 
-void MeshNode::Draw(const glm::mat4& topMatrix, DrawContext& ctx) {
-    glm::mat4 nodeMatrix = topMatrix * worldTransform;
-
-    for (auto& s : mesh->surfaces) {
-        RenderObject def{};
-        def.indexCount = s.count;
-        def.firstIndex = s.startIndex;
-        def.indexBuffer = mesh->meshBuffers.indexBuffer.buffer;
-        def.material = &s.material->data;
-
-        def.transform = nodeMatrix;
-        def.vertexBufferAddress = mesh->meshBuffers.vertexBufferAddress;
-
-        ctx.OpaqueSurfaces.push_back(def);
-    }
-
-    ENode::Draw(topMatrix, ctx);
-}
-
 void VulkanEngine::update_scene(const IModel::Ptr& model) {
     mainCamera->update();
